move o.a.l.document to o.a.l.sandbox.document; relax visibility of o.a.l.RangeFieldQuery
