+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/document/package.html
@@ -25,7 +25,7 @@
 <body>
 This package contains several point types:
 <ul>
-   <li>{@link org.apache.lucene.document.BigIntegerPoint BigIntegerPoint} for 128-bit integers</li>
+   <li>{@link org.apache.lucene.sandbox.document.BigIntegerPoint BigIntegerPoint} for 128-bit integers</li>
    <li>{@link org.apache.lucene.document.LatLonPoint LatLonPoint} for latitude/longitude geospatial points</li>
 </ul>
 </body>
