+++ lucene/sandbox/src/test/org/apache/lucene/sandbox/document/TestDoubleRangeField.java
@@ -14,8 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.document;
+package org.apache.lucene.sandbox.document;
 
+import org.apache.lucene.document.Document;
+import org.apache.lucene.document.DoubleRange;
 import org.apache.lucene.util.LuceneTestCase;
 
 /**
